requirements.txt
pytest
requests
python-dotenv
pytest-rerunfailures
pytest-pyreport
psycopg2-binary
allure-pytest

